pyproject.toml
[tool.poetry]
name = "tgportal"
version = "0.1.0"
description = "FastAPI backend for Telegram Portal"
authors = ["Admin <admin@example.com>"]
readme = "README.md"
packages = [{include = "server"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.12"
uvicorn = { extras = ["standard"], version = "^0.29" }
python-telegram-bot = "^20.6"
pydantic = "^2.5.2"
python-dotenv = "^1.0.0"
httpx = "^0.25.2"
telethon = "^1.40.0"
sqlalchemy = "^2.0.20"
alembic = "^1.12.4"
pydantic-settings = "^2.0.3"
psycopg2-binary = "^2.9.10"
sqlalchemy-utils = "^0.41.2"
asyncpg = "^0.30.0"
dotenv = "^0.9.9"
greenlet = "^3.2.3"
loguru = "^0.7.3"
google-generativeai = "^0.8.5"

[tool.poetry.scripts]
app = "server.app.main:run"
serve = "uvicorn server.app.main:app --reload"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
